rename coef_function_factory.h/cpp
