+++ tests/core/calibration/CMakeLists.txt
@@ -3,7 +3,7 @@ set(dir ${CMAKE_CURRENT_SOURCE_DIR})
 set(SOURCES
   ${dir}/parameter.cpp
   ${dir}/coef_function.cpp
-  ${dir}/coef_function_factory.cpp
+  ${dir}/CoefFunctionFactoryTest.cpp
   ${dir}/PolynomialTest.cpp
   ${dir}/CalibrationTest.cpp
   )
